refactor: split stats analytics helpers
